backend/ml/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
tqdm>=4.0
# ultralytics is optional (YOLOv8) for segmentation training
ultralytics>=8.0.0
